% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.packages.ifmissing.R
\name{install.packages.ifmissing}
\alias{install.packages.ifmissing}
\title{Install Packages (If Missing)}
\usage{
install.packages.ifmissing(pkgs, lib = .libPaths(), ...)
}
\arguments{
\item{pkgs}{A character vector of the names of packages which should be installed if they are not installed already.}

\item{lib}{A character vector giving the library directories where to look for and install the packages. If missing, defaults to the first element of \code{\link{.libPaths()}}.}

\item{...}{Additional arguments to be passed to \code{\link{install.packages}}.}
}
\description{
Installs packages that have not already been installed.
}
\details{
Installs packages that have not already been installed.

Packages are not updated with the latest version if an older version has been installed.
}
\examples{
# Installs the dplyr package:
install.packages.ifmissing("dplyr")

# Installs the ggplot2 package, but
# will not reinstall the dplyr package.
install.packages.ifmissing(c("dplyr", "ggplot2"))

}
\seealso{
\code{\link{installed.packages}} which is used to determine which packages have already been installed.

\code{\link{install.packages}} which is used to install packages.
}

