% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_output_folder.R
\name{project_output_folder}
\alias{project_output_folder}
\alias{project_output_folder_path}
\title{Project Output Folder}
\usage{
project_output_folder(folder = NULL)

project_output_folder_path()
}
\description{
Get or set the current output folder.
}

