% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_libraries.R
\name{project_packages}
\alias{project_packages}
\title{Project packages}
\usage{
project_packages(pkgs)
}
\arguments{
\item{pkgs}{A character vector with the required packages.}
}
\description{
Installs packages if missing and loads the libraries
}
\details{
Installs packages if missing and loads the libraries
}
\examples{
project_packages("dplyr")
project_packages(c("dplyr", "ggplot2"))

}
\seealso{
\code{\link{install.packages.ifmissing}}
\code{\link{library}}
}

