% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config_reload.R
\name{project_reload_config}
\alias{project_reload_config}
\title{Reload Project Config}
\usage{
project_reload_config(config_file = NULL, keepDTS = FALSE)
}
\arguments{
\item{config_file}{A string specifying the path to the configuration file. If not specified, the \code{config.yml} file in the working directory will be loaded if it exists, otherwise a default configuration file will be loaded from the package.}

\item{keepDTS}{A boolean flag indicating whether the previously set DateTime Stamp should be reused (if \code{project_reload_config} has been used before to set \code{project_config$DTS}).}
}
\description{
(Re)loads the project configuration file.
}
\details{
Loads or reloads the project configuration file.

The previous DateTime Stamp will be preserved is so indicated.
}
\examples{
project_config_reload()
project_config_reload("my_config.yml")

}
\seealso{
\code{\link{project_restart}}.
}

