% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_restart.R
\name{project_restart}
\alias{project_restart}
\title{Restart a project}
\usage{
project_restart(lib = NULL, config_file = NULL)
}
\arguments{
\item{lib}{A character vector with the locations of R library trees.}

\item{config_file}{A string specifying the path to the configuration file. If not specified, the \code{config.yml} file in the working directory will be loaded if it exists, otherwise a default configuration file will be loaded from the package.}
}
\description{
(Re)starts the project.
}
\details{
(Re)starts the project by performing the following steps:
\itemize{
\item Clears the environment.
\item Sets \code{.libPaths} if specified.
\item Loads the project configuration file into an R variable \code{project_config}.
\item Ensures all the folders specified in the configuration file exist.
\item Initialises logging.
\item Initialises the models list.
}
}
\examples{
project_restart()
project_restart("my_config.yml")

}
\seealso{
\code{\link{.libPaths}}

\code{\link{project_config_reload}}

\code{\link{ensure_folder}}

\code{\link{logging}}
}

