% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_section.R
\name{project_section}
\alias{project_section}
\alias{project_subsection}
\title{Project Section}
\usage{
project_section(name = NULL, folder = NULL)

project_subsection(name, folder = NULL)
}
\description{
Starts a project section.
}

