% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim.cached.R
\name{read_delim.cached}
\alias{read_csv.cached}
\alias{read_csv2.cached}
\alias{read_delim.cached}
\alias{read_tsv.cached}
\title{Read a delimited file into a data frame (Cached)}
\usage{
read_delim.cached(rds, file, ...)

read_csv.cached(rds, file, ...)

read_csv2.cached(rds, file, ...)

read_tsv.cached(rds, file, ...)
}
\arguments{
\item{rds}{The name of the file in which the cache is to be stored.}

\item{file}{The name of the file which the data are to be read from.}

\item{...}{Additional arguments to be passed to the base function.}
}
\description{
Wrapper functions to implement caching for \code{\link{read_delim}}-based functions from the \code{readr} package using RDS-files.
}
\details{
If the \code{rds} file exists, the object will be loaded from the \code{rds}.
Otherwise the \code{file} will be read and an \code{rds} of the resulting
object will be saved.
}
\examples{
read_csv.cached("testdata.rds", "data/testdata.csv")

}
\seealso{
\code{\link{do.call.cached}} which is the wrapper function that performs the caching.

\code{\link{read_delim}} which is called to read the delimited data file.
}

