% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_evenly.R
\name{split_evenly}
\alias{split_evenly}
\title{Split a vector in equal groups}
\usage{
split_evenly(x, n)
}
\arguments{
\item{x}{A vector containing values to be split into groups.}

\item{n}{The number of groups to create.}
}
\description{
Splits a vector into groups of (almost) equal length.
}
\examples{
split_evenly(1:10, 4)

}

