% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logging.R
\name{start_logging}
\alias{start_logging}
\title{Start Logging}
\usage{
start_logging(level = "INFO", file = NULL, logger = "", reset = TRUE)
}
\arguments{
\item{level}{The logging level of the root logger. Defaults to "INFO".}

\item{file}{Optional file name or path for logging to a file. Defaults to NULL.}

\item{logger}{Character string indicating the logger for the handler. Defaults to the root logger.}

\item{reset}{Boolean value indicating whether to reset the logging to the \code{basicConfig}.}
}
\description{
Initialises the logging to the console and optionally a file.
}
\details{
Starts logging by loading the \code{logging} package, resetting the logging
(reset = TRUE, the default). Then it sets up a handler to log to the console,
and if a file has been specified it sets up a handler to log to the file.
Finally a message is logged at the INFO level to indicate that the logging has been started.
}
\examples{
start_logging()
start_logging(file = 'log.txt')
start_logging(  level = 'CRITICAL'
              , logger = 'my_logger'
              , reset = FALSE)

}
\seealso{
\code{\link{logging}} a tentative logging package
}

